# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Zuultimate
  version: 1.0.0
paths:
...
  /v1/sso/providers:
    post:
      tags:
      - sso
      summary: Create SSO provider
      operationId: create_provider_v1_sso_providers_post
      security:
      - HTTPBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SSOProviderCreate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SSOProviderResponse'
        '401':
          description: Authentication required or token invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      x-stability: stable
      x-required-scopes:
      - identity:write
    get:
      tags:
      - sso
      summary: List SSO providers
      operationId: list_providers_v1_sso_providers_get
      security:
      - HTTPBearer: []
      parameters:
      - name: tenant_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Tenant Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SSOProviderResponse'
                title: Response List Providers V1 Sso Providers Get
        '401':
          description: Authentication required or token invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      x-stability: stable
      x-required-scopes:
      - identity:read
  /v1/sso/login/{provider_id}:
    get:
      tags:
      - sso
      summary: Initiate SSO login
      operationId: initiate_sso_login_v1_sso_login__provider_id__get
      parameters:
      - name: provider_id
        in: path
        required: true
        schema:
          type: string
          title: Provider Id
      - name: redirect_uri
        in: query
        required: false
        schema:
          type: string
          default: http://localhost:3000/callback
          title: Redirect Uri
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SSOLoginInitResponse'
        '401':
          description: Authentication required or token invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      x-stability: stable
  /v1/sso/callback:
    post:
      tags:
      - sso
      summary: Handle SSO callback
      operationId: sso_callback_v1_sso_callback_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SSOCallbackRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '401':
          description: Authentication required or token invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      x-stability: stable
  /v1/sso/providers/{provider_id}:
    delete:
      tags:
      - sso
      summary: Deactivate SSO provider
      operationId: deactivate_provider_v1_sso_providers__provider_id__delete
      security:
      - HTTPBearer: []
      parameters:
      - name: provider_id
        in: path
        required: true
        schema:
          type: string
          title: Provider Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '401':
          description: Authentication required or token invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      x-stability: stable
      x-required-scopes:
      - identity:write
...
components:
  schemas:
...
    ErrorResponse:
      properties:
        error:
          type: string
          title: Error
          description: Error message
        code:
          type: string
          title: Code
          description: Machine-readable error code
          default: ZUUL_ERROR
        detail:
          anyOf:
          - type: string
          - type: 'null'
          title: Detail
          description: Additional error context
      type: object
      required:
      - error
      title: ErrorResponse
...
    SSOCallbackRequest:
      properties:
        provider_id:
          type: string
          title: Provider Id
          description: SSO provider UUID
        code:
          type: string
          title: Code
          description: Authorization code from IdP
        state:
          type: string
          title: State
          description: CSRF state parameter to verify
        nonce:
          type: string
          title: Nonce
          description: OIDC nonce to validate against stored value
        redirect_uri:
          type: string
          title: Redirect Uri
          description: Redirect URI used in authorization request
          default: ''
      type: object
      required:
      - provider_id
      - code
      - state
      - nonce
      title: SSOCallbackRequest
    SSOLoginInitResponse:
      properties:
        redirect_url:
          type: string
          title: Redirect Url
          description: URL to redirect the user to for authentication
        state:
          type: string
          title: State
          description: CSRF state parameter
        nonce:
          type: string
          title: Nonce
          description: OIDC nonce for replay protection
        provider_id:
          type: string
          title: Provider Id
          description: SSO provider UUID
      type: object
      required:
      - redirect_url
      - state
      - nonce
      - provider_id
      title: SSOLoginInitResponse
    SSOProviderCreate:
      properties:
        name:
          type: string
          maxLength: 255
          minLength: 1
          title: Name
          description: Provider display name
          examples:
          - Okta Production
        protocol:
          type: string
          pattern: ^oidc$
          title: Protocol
          description: SSO protocol (only 'oidc' supported)
          examples:
          - oidc
        issuer_url:
          type: string
          maxLength: 500
          minLength: 1
          title: Issuer Url
          description: Identity provider issuer URL
          examples:
          - https://accounts.google.com
        client_id:
          type: string
          maxLength: 255
          minLength: 1
          title: Client Id
          description: OAuth2 client ID
          examples:
          - my-client-id
        client_secret:
          type: string
          title: Client Secret
          description: OAuth2 client secret
          default: ''
        metadata_url:
          type: string
          title: Metadata Url
          description: OIDC discovery metadata URL
          default: ''
        tenant_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Tenant Id
          description: Scope provider to a specific tenant
      type: object
      required:
      - name
      - protocol
      - issuer_url
      - client_id
      title: SSOProviderCreate
    SSOProviderResponse:
      properties:
        id:
          type: string
          title: Id
          description: Provider UUID
        name:
          type: string
          title: Name
          description: Provider display name
        protocol:
          type: string
          title: Protocol
          description: SSO protocol
        issuer_url:
          type: string
          title: Issuer Url
          description: Identity provider issuer URL
        client_id:
          type: string
          title: Client Id
          description: OAuth2 client ID
        metadata_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Metadata Url
          description: OIDC discovery URL
        tenant_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Tenant Id
          description: Associated tenant ID
        is_active:
          type: boolean
          title: Is Active
          description: Whether provider is active
      type: object
      required:
      - id
      - name
      - protocol
      - issuer_url
      - client_id
      - is_active
      title: SSOProviderResponse
...
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer